provide an ability to search for best config

+++ ConfigurationAnalyzer.Domain/Models/Run/ConfigurationRunPropertiesProcessed.cs
@@ -2,13 +2,13 @@
 {
 	public class ConfigurationRunPropertiesProcessed: ConfigurationRunProperties
 	{
-		public double DurationBest { get; set; }
+		public int DurationBest { get; set; }
 
-		public double DurationWorst { get; set; }
+		public int DurationWorst { get; set; }
 
-		public double InefficiencyTimeBest { get; set; }
+		public int InefficiencyTimeBest { get; set; }
 
-		public double InefficiencyTimeWorst { get; set; }
+		public int InefficiencyTimeWorst { get; set; }
 
 
 	}
